Use the same event subscription model in both STDIO API and WS API
